initialize game and board

+++ src/main/java/com/yergun/bol/mancala/model/Score.java
@@ -11,6 +11,6 @@ import javax.persistence.Embeddable;
 @NoArgsConstructor
 @Embeddable
 public class Score {
-    private Long playerOneScore;
-    private Long playerTwoScore;
+    private Integer playerOneScore;
+    private Integer playerTwoScore;
 }
